Reintroduce the EventFlow.MsSql project... lol full circle...
