+++ EventFlow.sln
@@ -70,6 +70,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SQL", "SQL", "{92F3C263-8C0
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventFlow.Sql", "Source\EventFlow.Sql\EventFlow.Sql.csproj", "{8EC9D877-7351-4C7E-87A8-29BBD3D7FEBB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventFlow.MsSql", "Source\EventFlow.MsSql\EventFlow.MsSql.csproj", "{A0C8E77F-DB74-4035-B0DE-9BADDFCA5EBE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -164,6 +166,10 @@ Global
 		{8EC9D877-7351-4C7E-87A8-29BBD3D7FEBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8EC9D877-7351-4C7E-87A8-29BBD3D7FEBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8EC9D877-7351-4C7E-87A8-29BBD3D7FEBB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A0C8E77F-DB74-4035-B0DE-9BADDFCA5EBE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A0C8E77F-DB74-4035-B0DE-9BADDFCA5EBE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A0C8E77F-DB74-4035-B0DE-9BADDFCA5EBE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A0C8E77F-DB74-4035-B0DE-9BADDFCA5EBE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -189,5 +195,6 @@ Global
 		{079DC9A6-7740-476A-BC4F-D74D8D7A32AC} = {7F20CC09-6988-40BC-8693-80E0F7022926}
 		{A6F9C84B-27F6-4179-9768-4AA2099405FC} = {7F20CC09-6988-40BC-8693-80E0F7022926}
 		{8EC9D877-7351-4C7E-87A8-29BBD3D7FEBB} = {92F3C263-8C0C-4D12-B41A-452E48D2E5E8}
+		{A0C8E77F-DB74-4035-B0DE-9BADDFCA5EBE} = {E4FC24C0-3EB3-4203-B4F2-0B534B42574A}
 	EndGlobalSection
 EndGlobal
