+++ Source/EventFlow.EventStores.MsSql/EventFlow.EventStores.MsSql.csproj
@@ -50,6 +50,10 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\EventFlow.MsSql\EventFlow.MsSql.csproj">
+      <Project>{a0c8e77f-db74-4035-b0de-9baddfca5ebe}</Project>
+      <Name>EventFlow.MsSql</Name>
+    </ProjectReference>
     <ProjectReference Include="..\EventFlow.Sql\EventFlow.Sql.csproj">
       <Project>{8ec9d877-7351-4c7e-87a8-29bbd3d7febb}</Project>
       <Name>EventFlow.Sql</Name>
