+++ Source/EventFlow.MsSql.Tests/EventFlow.MsSql.Tests.csproj
@@ -61,6 +61,10 @@
       <Project>{0c92026e-7e6d-40b9-a4dd-a7041e23aa7d}</Project>
       <Name>EventFlow.EventStores.MsSql</Name>
     </ProjectReference>
+    <ProjectReference Include="..\EventFlow.MsSql\EventFlow.MsSql.csproj">
+      <Project>{a0c8e77f-db74-4035-b0de-9baddfca5ebe}</Project>
+      <Name>EventFlow.MsSql</Name>
+    </ProjectReference>
     <ProjectReference Include="..\EventFlow.ReadStores.MsSql\EventFlow.ReadStores.MsSql.csproj">
       <Project>{cc886dd8-a471-4a08-822b-46f6dc5ae347}</Project>
       <Name>EventFlow.ReadStores.MsSql</Name>
