+++ Source/EventFlow.MsSql.Tests/IntegrationTests/EventStores/MsSqlEventStoreTests.cs
@@ -25,8 +25,8 @@
 using EventFlow.Configuration;
 using EventFlow.EventStores.MsSql;
 using EventFlow.Extensions;
+using EventFlow.MsSql.Extensions;
 using EventFlow.Sql;
-using EventFlow.Sql.Extensions;
 using EventFlow.TestHelpers;
 using EventFlow.TestHelpers.Suites;
 using Helpz.MsSql;
