+++ Source/EventFlow.MsSql.Tests/IntegrationTests/ReadStores/MsSqlReadModelStoreTests.cs
@@ -27,11 +27,11 @@ using System.Threading.Tasks;
 using EventFlow.Configuration;
 using EventFlow.EventStores.MsSql;
 using EventFlow.Extensions;
+using EventFlow.MsSql.Extensions;
 using EventFlow.MsSql.Tests.IntegrationTests.ReadStores.QueryHandlers;
 using EventFlow.MsSql.Tests.IntegrationTests.ReadStores.ReadModels;
 using EventFlow.ReadStores.MsSql.Extensions;
 using EventFlow.Sql;
-using EventFlow.Sql.Extensions;
 using EventFlow.Sql.Migrations;
 using EventFlow.TestHelpers;
 using EventFlow.TestHelpers.Aggregates.Entities;
