+++ Source/EventFlow.MsSql/EventFlow.MsSql.csproj
@@ -1,14 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{CC886DD8-A471-4A08-822B-46F6DC5AE347}</ProjectGuid>
+    <ProjectGuid>{A0C8E77F-DB74-4035-B0DE-9BADDFCA5EBE}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>EventFlow.ReadStores.MsSql</RootNamespace>
-    <AssemblyName>EventFlow.ReadStores.MsSql</AssemblyName>
+    <RootNamespace>EventFlow.MsSql</RootNamespace>
+    <AssemblyName>EventFlow.MsSql</AssemblyName>
     <TargetFrameworkVersion>v4.5.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -20,7 +20,6 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
@@ -29,39 +28,30 @@
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
-    <Reference Include="System.ComponentModel.DataAnnotations" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Data" />
+    <Reference Include="System.Net.Http" />
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\SolutionInfo.cs">
-      <Link>Properties\SolutionInfo.cs</Link>
-    </Compile>
-    <Compile Include="Attributes\MsSqlReadModelIdentityColumnAttribute.cs" />
-    <Compile Include="Attributes\MsSqlReadModelIgnoreColumnAttribute.cs" />
-    <Compile Include="Attributes\MsSqlReadModelVersionColumnAttribute.cs" />
     <Compile Include="Extensions\EventFlowOptionsExtensions.cs" />
-    <Compile Include="IMssqlReadModel.cs" />
-    <Compile Include="IMssqlReadModelStore.cs" />
-    <Compile Include="IReadModelSqlGenerator.cs" />
-    <Compile Include="MssqlReadModel.cs" />
-    <Compile Include="MssqlReadModelStore.cs" />
+    <Compile Include="IMsSqlConfiguration.cs" />
+    <Compile Include="IMsSqlConnection.cs" />
+    <Compile Include="IMsSqlDatabaseMigrator.cs" />
+    <Compile Include="MsSqlConfiguration.cs" />
+    <Compile Include="MsSqlConnection.cs" />
+    <Compile Include="MsSqlDatabaseMigrator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ReadModelSqlGenerator.cs" />
+    <Compile Include="RetryStrategies\IMsSqlErrorRetryStrategy.cs" />
+    <Compile Include="RetryStrategies\MsSqlErrorRetryStrategy.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\EventFlow.MsSql\EventFlow.MsSql.csproj">
-      <Project>{a0c8e77f-db74-4035-b0de-9baddfca5ebe}</Project>
-      <Name>EventFlow.MsSql</Name>
-    </ProjectReference>
     <ProjectReference Include="..\EventFlow.Sql\EventFlow.Sql.csproj">
       <Project>{8ec9d877-7351-4c7e-87a8-29bbd3d7febb}</Project>
       <Name>EventFlow.Sql</Name>
@@ -72,8 +62,7 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <None Include="EventFlow.ReadStores.MsSql.nuspec" />
-    <None Include="paket.references" />
+    <None Include="EventFlow.MsSql.nuspec" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
@@ -83,16 +72,4 @@
   <Target Name="AfterBuild">
   </Target>
   -->
-  <Import Project="..\..\.paket\paket.targets" />
-  <Choose>
-    <When Condition="$(TargetFrameworkIdentifier) == '.NETFramework' And ($(TargetFrameworkVersion) == 'v4.5.1' Or $(TargetFrameworkVersion) == 'v4.5.2' Or $(TargetFrameworkVersion) == 'v4.5.3' Or $(TargetFrameworkVersion) == 'v4.6' Or $(TargetFrameworkVersion) == 'v4.6.1')">
-      <ItemGroup>
-        <Reference Include="Newtonsoft.Json">
-          <HintPath>..\..\packages\Newtonsoft.Json\lib\net45\Newtonsoft.Json.dll</HintPath>
-          <Private>True</Private>
-          <Paket>True</Paket>
-        </Reference>
-      </ItemGroup>
-    </When>
-  </Choose>
 </Project>
