+++ Source/EventFlow.MsSql/IMsSqlConnection.cs
@@ -22,26 +22,11 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-using System;
-using EventFlow.Core;
 using EventFlow.Sql.Connections;
 
-namespace EventFlow.Sql
+namespace EventFlow.MsSql
 {
-    public class MsSqlConfiguration : SqlConfiguration<IMsSqlConfiguration>, IMsSqlConfiguration
+    public interface IMsSqlConnection : ISqlConnection
     {
-        public static MsSqlConfiguration New => new MsSqlConfiguration();
-
-        public RetryDelay TransientRetryDelay { get; private set; } = RetryDelay.Between(
-            TimeSpan.FromMilliseconds(50),
-            TimeSpan.FromMilliseconds(100));
-
-        private MsSqlConfiguration() { }
-
-        public IMsSqlConfiguration SetTransientRetryDelay(RetryDelay retryDelay)
-        {
-            TransientRetryDelay = retryDelay;
-            return this;
-        }
     }
 }
