+++ Source/EventFlow.MsSql/IMsSqlDatabaseMigrator.cs
@@ -20,17 +20,13 @@
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-//
+// 
 
-using EventFlow.Core;
-using EventFlow.Sql.Connections;
+using EventFlow.Sql.Migrations;
 
-namespace EventFlow.Sql
+namespace EventFlow.MsSql
 {
-    public interface IMsSqlConfiguration : ISqlConfiguration
+    public interface IMsSqlDatabaseMigrator : ISqlDatabaseMigrator
     {
-        RetryDelay TransientRetryDelay { get; }
-
-        IMsSqlConfiguration SetTransientRetryDelay(RetryDelay retryDelay);
     }
 }
