+++ Source/EventFlow.MsSql/MsSqlDatabaseMigrator.cs
@@ -20,20 +20,19 @@
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-//
+// 
 
-using EventFlow.Core;
-using EventFlow.Sql.Connections;
-using EventFlow.Sql.RetryStrategies;
+using EventFlow.Logs;
+using EventFlow.Sql.Migrations;
 
-namespace EventFlow.Sql
+namespace EventFlow.MsSql
 {
-    public class MsSqlConnection : SqlConnection<IMsSqlConfiguration, IMsSqlErrorRetryStrategy>, IMsSqlConnection
+    public class MsSqlDatabaseMigrator : SqlDatabaseMigrator<IMsSqlConfiguration>, IMsSqlDatabaseMigrator
     {
-        public MsSqlConnection(
-            IMsSqlConfiguration configuration,
-            ITransientFaultHandler<IMsSqlErrorRetryStrategy> transientFaultHandler)
-            : base(configuration, transientFaultHandler)
+        public MsSqlDatabaseMigrator(
+            ILog log,
+            IMsSqlConfiguration sqlConfiguration)
+            : base(log, sqlConfiguration)
         {
         }
     }
