+++ Source/EventFlow.ReadStores.MsSql/MssqlReadModelStore.cs
@@ -32,6 +32,7 @@ using EventFlow.Aggregates;
 using EventFlow.Core;
 using EventFlow.Extensions;
 using EventFlow.Logs;
+using EventFlow.MsSql;
 using EventFlow.ReadStores.MsSql.Attributes;
 using EventFlow.Sql;
